+++ modules/tinymce/src/core/test/ts/browser/init/EditorInitializationTest.ts
@@ -256,4 +256,8 @@ describe('browser.tinymce.core.init.EditorInitializationTest', () => {
   it('Initialize on list item with initial content', (done) => {
     initAndAssertContent('<ul><li>Initial Content</li></ul>', 'li', 'Initial Content', done);
   });
+
+  it('TINY-10337: Should remove ZWNBSP from content when initializing', (done) => {
+    initAndAssertContent('<textarea>te\uFEFFst</textarea>', 'textarea', 'test', done);
+  });
 });